+++ classes/plugin/PaginationTwigExtension.php
@@ -1,7 +1,8 @@
 <?php
-namespace Grav\Plugin;
+namespace Grav\Plugin\Pagination;
 
 use \Grav\Common\Grav;
+use Grav\Plugin\PaginationPlugin;
 
 class PaginationTwigExtension extends \Twig_Extension
 {
@@ -35,7 +36,8 @@ class PaginationTwigExtension extends \Twig_Extension
         ];
     }
 
-    public function paginateFunc($collection, $limit, $ignore_url_param_array = []) {
+    public function paginateFunc($collection, $limit, $ignore_url_param_array = [])
+    {
         $pag = new PaginationPlugin($this, Grav::instance(), $this->config);
         $pag->paginateCollection($collection, $limit, $ignore_url_param_array);
     }