+++ pagination.php
@@ -1,9 +1,13 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
 use Grav\Common\Page\Collection;
 use Grav\Common\Page\Interfaces\PageInterface;
 use Grav\Common\Plugin;
+use Grav\Plugin\Pagination\PaginationHelper;
+use Grav\Plugin\Pagination\PaginationPage;
+use Grav\Plugin\Pagination\PaginationTwigExtension;
 use RocketTheme\Toolbox\Event\Event;
 
 class PaginationPlugin extends Plugin
@@ -19,8 +23,21 @@ class PaginationPlugin extends Plugin
     public static function getSubscribedEvents()
     {
         return [
-            'onPluginsInitialized' => ['onPluginsInitialized', 0]
+            'onPluginsInitialized' => [
+                ['autoload', 100001],
+                ['onPluginsInitialized', 0]
+            ]
         ];
+    }
+
+    /**
+     * [onPluginsInitialized:100000] Composer autoload.
+     *
+     * @return ClassLoader
+     */
+    public function autoload()
+    {
+        return require __DIR__ . '/vendor/autoload.php';
     }
 
     /**
@@ -32,6 +49,10 @@ class PaginationPlugin extends Plugin
             $this->active = false;
             return;
         }
+
+        class_alias(PaginationHelper::class, 'Grav\\Plugin\\PaginationHelper');
+        class_alias(PaginationPage::class, 'Grav\\Plugin\\PaginationPage');
+        class_alias(PaginationTwigExtension::class, 'Grav\\Plugin\\PaginationTwigExtension');
 
         $this->enable([
             'onTwigTemplatePaths' => ['onTwigTemplatePaths', 0],
@@ -53,8 +74,6 @@ class PaginationPlugin extends Plugin
      */
     public function onTwigExtensions()
     {
-        require_once(__DIR__.'/twig/PaginationTwigExtension.php');
-
         $this->grav['twig']->twig->addExtension(new PaginationTwigExtension());
     }
 
@@ -96,7 +115,6 @@ class PaginationPlugin extends Plugin
         }
 
         if (!empty($params['limit']) && $collection->count() > $params['limit']) {
-            require_once __DIR__ . '/classes/paginationhelper.php';
             $this->pagination = new PaginationHelper($collection);
             $collection->setParams(['pagination' => $this->pagination]);
         }
@@ -119,14 +137,13 @@ class PaginationPlugin extends Plugin
      * @param int $limit
      * @param array $ignore_param_array      url parameters to be ignored in page links
      */
-    public function paginateCollection( $collection, $limit, $ignore_param_array = [])
+    public function paginateCollection($collection, $limit, $ignore_param_array = [])
     {
         $collection->setParams(['pagination' => 'true']);
         $collection->setParams(['limit' => $limit]);
         $collection->setParams(['ignore_params' => $ignore_param_array]);
 
         if ($collection->count() > $limit) {
-            require_once __DIR__ . '/classes/paginationhelper.php';
             $this->pagination = new PaginationHelper($collection);
             $collection->setParams(['pagination' => $this->pagination]);