+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.4.3
+## mm/dd/2019
+
+1. [](#new)
+    * Pass phpstan level 1 tests
+    * Require Grav v1.6
+    * Moved pagination classes under `Grav\Plugin\Pagination` namespace
+
 # v1.4.2
 ## 05/09/2019